Create LICENSE.md
This commit adds the Creative Commons Attribution 4.0 International License (CC-BY-4.0) to the repository, providing clear guidelines for sharing and adapting the content. This license ensures that contributors and users understand their rights and responsibilities regarding the material shared within the repository. By incorporating the CC-BY-4.0 license, we promote a culture of collaboration and openness while respecting the original author's copyright.