Update README.md
This commit revises the README file to include information about the Creative Commons Attribution 4.0 International License (CC-BY-4.0) for the repository. By adding this license information directly to the README, we ensure that users and contributors are aware of the licensing terms governing the content shared within the repository. This update promotes transparency and facilitates adherence to copyright and licensing requirements, fostering a collaborative environment for sharing and adapting the material.